refactor: remove 'wrapped' ref from parent component
This is causing a problem, for some reason, only on a class-style
component, where it's complaining that: index.js:327 Uncaught Error:
addComponentAsRefTo(...): Only a ReactOwner can have refs. You might be
adding a ref to a component that was not created inside a component's
`render` method, or you have multiple copies of React loaded (details:
https://fb.me/react-refs-must-have-owner).

Since only the tests, that I know of, relied on this anyway, I'm
removing it, since the test assertion that the parent is not the wrapped
component seems to be strong enough without this.

BREAKING CHANGE dunno if anyone relies on this.  Technically a part of
the public api